feat: add data storing in redux store
