+++ src/js/store/types.ts
@@ -1,0 +1,34 @@
+class User {
+  name: string;
+
+  city: string;
+
+  score: number;
+
+  isActive: boolean;
+
+  constructor(name, city, score, isActive) {
+    this.name = name;
+    this.city = city;
+    this.score = score;
+    this.isActive = isActive;
+  }
+
+  getName() {
+    return this.name;
+  }
+
+  getCity() {
+    return this.city;
+  }
+
+  getScore() {
+    return this.score.toLocaleString();
+  }
+
+  getIsActive() {
+    return this.isActive ? 'yes' : 'no';
+  }
+}
+
+export default User;
